% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19cv.R
\docType{data}
\name{covid19cv_cidades}
\alias{covid19cv_cidades}
\title{Data from Covid19 by cities in Cabo Verde.}
\format{
A data.frame with eleven variables: \code{Data}, \code{cidade},
  \code{lat}, \code{long} , \code{ilha},\code{recuperados},\code{obitos},
  \code{evacuados},\code{confirmados},\code{confirmados_acumulados},\code{c_ativos_acumulados}
}
\source{
Public government website-covid19.cv
}
\usage{
covid19cv_cidades
}
\description{
Dataset organized with type of cases by cities
}
\keyword{datasets}
