% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19cv.R
\docType{data}
\name{covid19cv_nacional}
\alias{covid19cv_nacional}
\title{Daily Summuray of Covid19 cases in Cabo Verde.}
\format{
A data.frame object with eleven variables: \code{Data}, \code{recuperados},
  \code{obitos}, \code{evacuados} , \code{confirmados}
}
\source{
Public government website-covid19.cv
}
\usage{
covid19cv_nacional
}
\description{
Dataset organized with type of cases by date
}
\keyword{datasets}
