% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19cv.R
\docType{data}
\name{covid19cv_pop}
\alias{covid19cv_pop}
\title{Demographics Data of Covid19 Case in Cabo Verde.}
\format{
A data.frame with eleven variables: \code{grupo_etario}, \code{sexo},
  \code{recuperados}, \code{obitos} , \code{evacuados},\code{confirmados},\code{confirmados_ativos}
}
\source{
Public government website-covid19.cv
}
\usage{
covid19cv_pop
}
\description{
Dataset organized with type of cases by age group and sex
}
\keyword{datasets}
