% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19cv.R
\docType{package}
\name{covid19cvdata-package}
\alias{covid19cvdata}
\alias{covid19cvdata-package}
\title{covid19cvdata: Dataset for Cabo Verde Covid19 Data}
\description{
Dataset gather from public sources. This package contains all registered cases of covid19 in Cabo Verde and functions used to gather and clean. Due the current situation of covid-19 in Cabo Verde there's been lacking 
        a easy way to access data about the numbers about the evolution of the pandemic in the country. 
        So, to address this issue I've manually gathered data from public sources (mainly the government website covid-19) and organized in csv files with a API to been able pass the information to any Data Nerd.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marovski/covid19cvdata}
  \item Report bugs at \url{https://github.com/marovski/covid19cvdata/issues}
}

}
\keyword{covid19cvdata}
