% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{data.update}
\alias{data.update}
\title{Update the Package datasets}
\usage{
data.update()
}
\description{
This function allows to update the datasets by comparing the changes in csv file
}
\examples{
\dontrun{
data.update()
}
}
\keyword{data.update()}
