% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{geo.reverse}
\alias{geo.reverse}
\title{Reverse lat and long into island long name}
\usage{
geo.reverse(x)
}
\arguments{
\item{x}{for the dataframe}
}
\value{
the altered dataframe with island name
}
\description{
This function allows to extract the island name with reverse geocode
}
\examples{
\dontrun{
geo.reverse(covid19cv)
}

}
\keyword{geo.reverse}
