% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{get.geocode}
\alias{get.geocode}
\title{Get lat and long of the places in Cabo Verde}
\usage{
get.geocode(x, place)
}
\arguments{
\item{x}{for dataframe}

\item{place}{argument to specify if this island or city (chr)}
}
\value{
the altered dataframe with long and lat columns
}
\description{
This function allows to obtain the lat and lon of places in openstreetmaps API
}
\examples{
\dontrun{
get.geocode(covid19cv,covid19cv$concelho)
}
}
\keyword{get.geocode}
