% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{get.ilha}
\alias{get.ilha}
\title{Method to extract siglum}
\usage{
get.ilha(l)
}
\arguments{
\item{l}{for dataframe specific column}
}
\value{
the altered dataframe
}
\description{
This function allows to extract the island siglum
}
\examples{
\dontrun{
lapply(covid19cv$local,get.ilha)
}

}
\keyword{get.ilha}
