% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{insert.case}
\alias{insert.case}
\title{Method to add new data}
\usage{
insert.case(x)
}
\arguments{
\item{x}{for dataframe}
}
\value{
the altered dataframe
}
\description{
This function allows to insert new row
}
\examples{
\dontrun{
insert.case(covid19cv)
}

}
