% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19cv.R
\docType{data}
\name{covid19cv}
\alias{covid19cv}
\title{Data from Covid19 Cases in Cabo Verde.}
\format{
A data frame with eleven variables: \code{data}, \code{nacionalidade},
  \code{faixa_etaria}, \code{sexo} , \code{tipo_caso},\code{tipo_transmissao},
  \code{cidade}, \code{grupo_etario},\code{ilha},\code{long},\code{lat}
}
\usage{
covid19cv
}
\description{
Dataset organized with data from covid19.cv
}
\keyword{datasets}
